+++ mosviz/loaders/__init__.py
@@ -1,7 +1,0 @@
-"""
-Data loaders for MOSViz
-"""
-
-from .jwst_loaders import *
-from .hst_loaders import *
-from .deimos_loaders import *