+++ mosviz/plugins/table_generator/__init__.py
@@ -1,2 +1,0 @@
-from .nirspec_table import *
-from .nirspec_table_ui import *