+++ mosviz/viewers/__init__.py
@@ -1,1 +1,0 @@
-from .mos_viewer import *