+++ mosviz/widgets/__init__.py
@@ -1,3 +1,0 @@
-from .plots import *
-from .toolbars import *
-from .viewer_options import *